This is what I've done thus far, and I'm not really going to sleep until i finish it but like i said i could use some skype guidance:
GroupProAct. :
Got the drop down menu and it shows the list of actions and i'm gonna
run the actions by doing a series of if statements in the
onItemSelected() that get which item was selected by the statement:
parent.getItemSelectedAt(pos).toString() == "(whatever the name of the
item in the drop-down menu is)" then inside the if statement call the
action. The calling of the actions is what I'm having trouble with.